Do not check /v2 endpoint on registry when RoundTripper is provided (#1396)
* Bump staticcheck to latest version

Refactor tests that where Marshalling the transport.Error. This was
causing staticcheck to fail.
Instead of trying to Marshall the error we now provide the strings
version of the expected errors.

Signed-off-by: João Pereira <joaod@vmware.com>

* Wrapper transport no longer checks /v2 endpoint

When the calling the function NewWithContext, if the RoundTripper
provided is of the type transport.Wrapper, the function will assume that
the caller did the needed due diligence to ensure the registry supports
OCI v2. In this case the function will only need to return back the
provided RoundTripper without any further change.

Signed-off-by: João Pereira <joaod@vmware.com>