Use the default retry predicate in transport (#1502)
We have two levels of retries here.

The first is in the retryTransport, which currently just tests
IsTemporary. This has a short backoff and was intended to handle blips
in the network.

The second level was intended for higher level issues, e.g. if an upload
PUT failed, we have to retry more than just a single HTTP request,
because we might have to restart the upload with a new POST and PATCH.

This change propagates the default retryable predicate (from the scond
level) down to the retryTransport. This also adds a list of HTTP codes
that are retryable, so that the retryTransport can also handle
server-side errors that should be retried.

Note that this doesn't respect remote.WithRetry{Backoff,Predicate}
because callers can just override the entire transport via
remote.WithTransport. This change tries to do the right thing by default
without changing behavior for folks who seem to already know what
they're doing.

Originally, I started down the path of wrapping various remote methods
in a retry.Retry function (similar to writes) and decided against it due
to the size of the change and how difficult it would be to maintain.

In order to reduce retry ampliciation, this adds a little context
wrapper to signal to retryTransport not to retry (because the retry is
happening at a higher level).

This also reduces the number of steps in the default backoff for
retryTransport because more kinds of things are being retried, and I
want to reduce the blast radius if something is erroneously retried.