Fix DNS lookup failure caused by too-short dial timeout
If the user's primary nameserver is unavailable, and they are using the default
resolver timeout of 5 seconds, the dial will fail before it can attempt
to use any backup nameservers.

This change raises the dial timeout to a value that should allow
failover through multiple nameservers. With a value of 16 seconds, the
4th nameserver has 1 second to respond before the dial times out.

Signed-off-by: Brad Davidson <brad@oatmail.org>