fix(mutate): also set timestamps only present in some formats (#1550)
Currently, layerTime only mutates the modification timestamp which is
present in all tar formats. When working with layers GNU or PAX header
formats, access and change timestamps are also generated/recorded. This
breaks functionality like reproducible builds which rely on having
control over all variable fields (esp. timestamps)  in the layers

This commit adds a special case to the layerTime(Layer, Time) function
which will also set the additional timestamps if the header is of GNU or
PAX format

Signed-off-by: Joel Pepper <pepper@bronze-deer.de>