Avoid v1.Manifest in crane edit config (#1583)
Because of how omitempty works, an empty Layers and a nil Layers will
both be omitted. We don't want that, because different registries
support different variants of that. Instead, we just pass through
whatever was there before by treating the manifest as a map[string]any
and manipulating that rather than using v1.Manifest-based helpers.