Make 403 non-fatal for manifest existence checks (#1691)
Pusher currently HEADs manifests to see if they already exist in the
destination repository. Unfortunately, Artifactory will return a 403 for
non-existent tags if you are using virtual repositories and do not have
access to one of the repositories backing the virtual repository. This
is almost certainly an issue with Artifactory, but it seems like this
was first reported in 2017, so this behavior may never change.