Don't wrap DefaultKeychain with refreshes (#1791)
The heuristic only works well if the upstream keychain doesn't do any
caching. This is usually the case, but I've run into a couple situations
where the hardcoded 5 minutes does not overlap well with certain cred
helper implementations.

Undoing that wrap allows callers to set a time that makes sense for
them. With the wrap, there's not any direct reference to
DefaultKeychain, which feels... bad.