name: Add tag.WithDigest.
Adds a new function to tags to allow appending a digest to the string
output. This allows users to output strings of the form
`repo:tag@digest`, which can often be used to display concise tag +
digest strings.

NOTE: Most clients parsing this string will ignore the tag in favor of the
digest (though there are a few exceptions like dependabot that will
place meaning in the tag). This is still useful for displaying resolved
tags to humans.