chore(deps): relax go directive to permit 1.22.x
Setting the go directive to 1.23.0 requires all consumers of the module
to be building with 1.23.0 or newer release of Go 1.23 and to update
their own go.mod accordingly — this seems unnecessarily restrictive for
a library module, particularly as the code itself doesn't currently use
any modern language constructs and builds fine even with older Go
versions.

Instead set the go directive to 1.22.0 and use the toolchain directive
to recommend the latest 1.23.x when building locally.

Note: this also includes a run of `hack/bump-deps.sh` to ensure all deps
are compatible with go1.22.0 as well

Fixes #2045

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>