Improve markdown for crane code snippets
Before this change, code snippets in the README were hard to copy since they contained a leading `$ ` and mixed shell commands with  console output. This change aims to improve the syntax so pure shell code snippets are easier to copy and so console output is marked as such (so GitHub can properly distinguish between command and output when rendering it). Also adds correct code fence for a YAML snippet.